Move middleware to folders, move UInt32 to UInt64

+++ Sources/BlockchainObserverInterface.swift
@@ -10,7 +10,7 @@ import Foundation
 
 public enum TransactionStatus {
     case registered
-    case confirmed(confirmations: UInt32)
+    case confirmed(confirmations: UInt64)
 }
 
 public protocol BlockchainObserverDelegate: class {
